add in local storage for users
